refactor(gantz_egui): split ops.rs into session and sync modules
Pure code motion: the undo/redo cluster (plain navigation ops, revert
commits, RevertCursor stepping) moves to ops/session.rs and the
merge/convergence cluster (apply_merge_migration, merge_head,
sync_remote_tip and their outcome types) to ops/sync.rs, each with its
tests. The edit ops, the layout/view helpers both clusters share
(node_id, cascade_pos, carry_layout, merged_view) and commit_layout stay
in the ops root, which re-exports the public items so external paths are
unchanged. Shared test fixtures (TestNode, test_graph,
diverged_registry) live in a cfg(test) ops::test_util.

Bonus tightening: revert_commit and apply_merge_migration become
module-private (each is called only within its new module).
